// src/models/IoK8sApiCoreV1NodeSystemInfo/types.ts
export interface IoK8sApiCoreV1NodeSystemInfoModelOptions {
  architecture: string;
  bootID: string;
  containerRuntimeVersion: string;
  kernelVersion: string;
  kubeProxyVersion: string;
  kubeletVersion: string;
  machineID: string;
  operatingSystem: string;
  osImage: string;
  systemUUID: string;
}